refactor(core): the client stops rebuilding a device it cannot see all of
Nothing reads the compatibility lane any more, so it goes. `AccountDialogSnapshot`
loses `accounts`, `activeAccountId` and `switchingAccountId`; the controller
loses `switchTo`, its `graph` and `profilesById` fields, and the two private
fetches behind them — `oxyServices.listAccounts()` and
`oxyServices.getUsersByIds()`. A dialog refresh is one request now, where it was
three.

The count is the smaller half of it. That reconstruction enumerated the CALLER's
account graph and presented the result as the device's, which on a device holding
two people is one person's answer wearing everyone's clothes — and it decided
switchability, an authorization question, on the client. Both are the server's
answers now, and `refresh()` reads the directory or nothing.

`commitSwitchedSession` and `locale` go with `switchTo`, which was the only
caller of the first and the only reason for the second, along with the provider's
funnel that fed it. `loading` re-keys onto the directory: a spinner is for having
nothing to show, and a re-read behind rows already on screen refreshes in place
rather than blanking them.

`accountProjection.ts` does NOT go. Two of its exports were never about the
device switcher — `canSwitchIntoAccount` and `isSwitchTargetAccount` ask a
question about the account GRAPH, which is what a caller can MANAGE and a longer
list than the identities this device can become — and they have live readers in
the Console's workspace tree and the Accounts app's managed-account rows. The
file keeps exactly those two and is renamed `accountSwitchTargets.ts`, because a
file called "projection" that no longer projects anything is a name that will
mislead the next reader.

BREAKING: `@oxyhq/core` removes `projectSwitchableAccounts`, `switchableAccountIds`,
`SwitchableAccount`, `SwitchableAccountUser`, `ProjectSwitchableAccountsInput`,
`AccountDialogSnapshot.{accounts,activeAccountId,switchingAccountId}`,
`AccountDialogController.switchTo`, and the `commitSwitchedSession` / `locale`
constructor options. A publicly exported symbol is public API even where the
org's own grep is empty — `core@13.2.0` cost us a deprecation and a republish for
exactly that — so this rides a major.

Core 117 suites / 1634 tests, services 76 / 688, auth 127, console 9 — green.

Refs #937
