+++ packages/core/src/session/__tests__/accountDialogController.test.ts
@@ -40,17 +40,38 @@ const EMPTY_DIRECTORY = {
   updatedAt: 1_720_000_000_000,
 };
 
-function host(): SessionClientHost {
+interface RecordingHost {
+  host: SessionClientHost;
+  /** Every path the client has requested, in order. */
+  urls: string[];
+  /** Make the next and every subsequent directory read reject with `error`. */
+  failDirectory: (error: unknown) => void;
+}
+
+function host(): RecordingHost {
+  const urls: string[] = [];
+  let directoryError: unknown = null;
   return {
-    makeRequest: jest.fn(async (_method: string, url: string) =>
-      (url === '/session/device/directory' ? EMPTY_DIRECTORY : undefined),
-    ),
-    getBaseURL: () => 'http://test.invalid',
-    getAccessToken: () => 'token',
-    getDeviceCredential: () => null,
-    onTokensChanged: () => () => undefined,
-    setTokens: jest.fn(),
-    getCurrentAccountId: () => null,
+    urls,
+    failDirectory: (error: unknown) => {
+      directoryError = error;
+    },
+    host: {
+      makeRequest: jest.fn(async (_method: string, url: string) => {
+        urls.push(url);
+        if (url === '/session/device/directory') {
+          if (directoryError !== null) throw directoryError;
+          return EMPTY_DIRECTORY;
+        }
+        return undefined;
+      }),
+      getBaseURL: () => 'http://test.invalid',
+      getAccessToken: () => 'token',
+      getDeviceCredential: () => null,
+      onTokensChanged: () => () => undefined,
+      setTokens: jest.fn(),
+      getCurrentAccountId: () => null,
+    },
   };
 }
 
@@ -153,6 +174,10 @@ interface Harness {
   sc: TestSessionClient;
   commitSession: jest.Mock;
   onSignedIn: jest.Mock;
+  /** Every path the session client has requested, in order. */
+  urls: string[];
+  /** Make every directory read reject with `error`. */
+  failDirectory: (error: unknown) => void;
 }
 
 function makeHarness(
@@ -166,7 +191,8 @@ function makeHarness(
   }> = {},
 ): Harness {
   const oxy = makeOxy();
-  const sc = new TestSessionClient(host());
+  const recording = host();
+  const sc = new TestSessionClient(recording.host);
   const commitSession = jest.fn().mockResolvedValue(undefined);
   const onSignedIn = jest.fn();
   const controller = createAccountDialogController({
@@ -182,18 +208,32 @@ function makeHarness(
     openUrl: over.openUrl,
     canOpenApp: over.canOpenApp,
   });
-  return { controller, oxy, sc, commitSession, onSignedIn };
+  return {
+    controller,
+    oxy,
+    sc,
+    commitSession,
+    onSignedIn,
+    urls: recording.urls,
+    failDirectory: recording.failDirectory,
+  };
 }
+
+/** How many times the device directory has been read. */
+const directoryReads = (urls: string[]): number =>
+  urls.filter((url) => url === '/session/device/directory').length;
 
 describe('AccountDialogController — initial + views', () => {
   it('starts on the accounts view with an empty list and idle sign-in', () => {
     const { controller } = makeHarness();
     const snap = controller.getSnapshot();
     expect(snap.view).toBe('accounts');
-    expect(snap.accounts).toEqual([]);
-    expect(snap.activeAccountId).toBeNull();
+    expect(snap.directory).toBeNull();
+    expect(snap.activeContext).toBeNull();
     expect(snap.loading).toBe(false);
-    expect(snap.switchingAccountId).toBeNull();
+    expect(snap.activatingContextId).toBeNull();
+    expect(snap.removingContextId).toBeNull();
+    expect(snap.removingPrincipalId).toBeNull();
     expect(snap.signIn.phase).toBe('idle');
     expect(snap.signIn.route).toBeNull();
     expect(snap.signIn.routeFailed).toBe(false);
@@ -240,314 +280,110 @@ describe('AccountDialogController — initial + views', () => {
   });
 });
 
-describe('AccountDialogController — account list', () => {
-  it('refresh loads graph + profiles and projects the unified list', async () => {
-    const { controller, oxy, sc } = makeHarness();
+describe('AccountDialogController — auth-gated directory read', () => {
+  it('start() while signed out makes no directory call and does not error', async () => {
+    const { controller, oxy, sc, urls } = makeHarness();
+    // Cold boot: no bearer planted yet (no listeners registered pre-start).
+    oxy.emitTokenChange(null);
     sc.set(state([{ accountId: 'a1', sessionId: 's1' }], 'a1'));
-    oxy.getUsersByIds.mockResolvedValue([user('a1'), user('org1')]);
-    oxy.listAccounts.mockResolvedValue([graphNode('org1')]);
 
-    await controller.refresh();
-
-    const snap = controller.getSnapshot();
-    expect(oxy.getUsersByIds).toHaveBeenCalledWith(['a1', 'org1']);
-    expect(snap.accounts.map((r) => r.accountId)).toEqual(['a1', 'org1']);
-    expect(snap.activeAccountId).toBe('a1');
-    expect(snap.accounts[0].isCurrent).toBe(true);
-    expect(snap.accounts[1].onDevice).toBe(false);
-    expect(snap.loading).toBe(false);
-  });
-
-  it('start subscribes to SessionClient so a device-state change re-projects', async () => {
-    const { controller, oxy, sc } = makeHarness();
-    oxy.getUsersByIds.mockResolvedValue([user('a1')]);
     controller.start();
-    await Promise.resolve();
-    await Promise.resolve();
+    await flush();
 
-    sc.set(state([{ accountId: 'a1', sessionId: 's1' }], 'a1'));
-    // The subscription re-projects synchronously from the new device state.
-    expect(controller.getSnapshot().activeAccountId).toBe('a1');
+    // The directory read is a PRIVATE call. Issuing it before the cold-boot
+    // restore plants the bearer 401s, which clears the token and signs the user
+    // out — the failure this gate exists for.
+    expect(urls).not.toContain('/session/device/directory');
+    const snap = controller.getSnapshot();
+    expect(snap.error).toBeNull();
+    expect(snap.loading).toBe(false);
     controller.destroy();
   });
 
-  it('keeps device rows and surfaces the error when listAccounts fails', async () => {
-    const warnSpy = jest.spyOn(logger, 'warn').mockImplementation(() => undefined);
+  it('start() while authenticated reads the directory exactly once', async () => {
+    const { controller, sc, urls } = makeHarness();
+    sc.set(state([{ accountId: 'a1', sessionId: 's1' }], 'a1'));
+
+    controller.start();
+    await flush();
+
+    expect(directoryReads(urls)).toBe(1);
+    controller.destroy();
+  });
+
+  it('reads the directory once when the bearer is planted after a signed-out start', async () => {
+    const { controller, oxy, urls } = makeHarness();
+    oxy.emitTokenChange(null);
+    controller.start();
+    await flush();
+    expect(urls).not.toContain('/session/device/directory');
+
+    // Cold-boot restore plants the token → onTokensChanged → one read.
+    oxy.emitTokenChange('access-token');
+    await flush();
+    expect(directoryReads(urls)).toBe(1);
+    controller.destroy();
+  });
+
+  it('never reconstructs the account graph — no listAccounts, no getUsersByIds', async () => {
     const { controller, oxy, sc } = makeHarness();
     sc.set(state([{ accountId: 'a1', sessionId: 's1' }], 'a1'));
-    oxy.getUsersByIds.mockResolvedValue([user('a1')]);
-    oxy.listAccounts.mockRejectedValue(new Error('graph boom'));
 
-    await controller.refresh();
+    controller.start();
+    await flush();
+    // A device change used to refetch profiles for any newly-seen account id.
+    sc.set(state([{ accountId: 'a1', sessionId: 's1' }, { accountId: 'a2', sessionId: 's2' }], 'a2', 2));
+    await flush();
 
-    const snap = controller.getSnapshot();
-    expect(snap.error).toBe('graph boom');
-    expect(snap.accounts.map((r) => r.accountId)).toEqual(['a1']);
-    // A genuine (non-401) failure STILL warns.
-    expect(warnSpy).toHaveBeenCalledWith(
-      '[AccountDialogController] listAccounts failed',
-      { component: 'AccountDialogController' },
-      expect.any(Error),
-    );
+    // ADR 0002's whole point: the client holds ONE caller's account graph and
+    // cannot enumerate another principal's, so a switcher assembled from it is
+    // one person's answer presented as the device's. The directory carries the
+    // rows AND the display metadata, so neither private call is made at all.
+    expect(oxy.listAccounts).not.toHaveBeenCalled();
+    expect(oxy.getUsersByIds).not.toHaveBeenCalled();
+    controller.destroy();
+  });
+
+  it('does not loop when the directory read rejects — one call per refresh, no re-trigger on device changes', async () => {
+    const warnSpy = jest.spyOn(logger, 'warn').mockImplementation(() => undefined);
+    const { controller, sc, urls, failDirectory } = makeHarness();
+    failDirectory(new Error('directory boom'));
+    sc.set(state([{ accountId: 'a1', sessionId: 's1' }], 'a1'));
+
+    controller.start();
+    await flush();
+    expect(directoryReads(urls)).toBe(1);
+    expect(controller.getSnapshot().error).toBe('directory boom');
+
+    // A subsequent device push must not re-trigger it (the auth edge did not
+    // move → `reconcileAuth` is a no-op → no storm).
+    sc.set(state([{ accountId: 'a1', sessionId: 's1' }], 'a1', 2));
+    await flush();
+    expect(directoryReads(urls)).toBe(1);
+    controller.destroy();
     warnSpy.mockRestore();
   });
 
-  it('treats a 401 from listAccounts as the signed-out edge — debug, no surfaced error, no warn', async () => {
+  it('treats a 401 from the directory read as the signed-out edge — debug, no surfaced error, no warn', async () => {
     const warnSpy = jest.spyOn(logger, 'warn').mockImplementation(() => undefined);
     const debugSpy = jest.spyOn(logger, 'debug').mockImplementation(() => undefined);
-    const { controller, oxy, sc } = makeHarness();
-    sc.set(state([{ accountId: 'a1', sessionId: 's1' }], 'a1'));
-    oxy.getUsersByIds.mockResolvedValue([user('a1')]);
-    // A stale/revoked bearer 401s: an EXPECTED signed-out outcome, not a failure.
-    oxy.listAccounts.mockRejectedValue(
+    const { controller, failDirectory } = makeHarness();
+    failDirectory(
       Object.assign(new Error('Invalid or missing authorization header'), { status: 401 }),
     );
 
     await controller.refresh();
 
-    const snap = controller.getSnapshot();
-    // Never surface an error for a normal signed-out state; device rows still render.
-    expect(snap.error).toBeNull();
-    expect(snap.accounts.map((r) => r.accountId)).toEqual(['a1']);
+    // A signed-out device is a normal state, not a warning.
+    expect(controller.getSnapshot().error).toBeNull();
     expect(warnSpy).not.toHaveBeenCalled();
     expect(debugSpy).toHaveBeenCalledWith(
-      '[AccountDialogController] listAccounts unauthorized (signed out)',
+      '[AccountDialogController] directory unauthorized (signed out)',
       { component: 'AccountDialogController' },
       expect.objectContaining({ status: 401 }),
     );
     warnSpy.mockRestore();
     debugSpy.mockRestore();
-  });
-});
-
-describe('AccountDialogController — auth-gated graph fetch (prod sign-out fix)', () => {
-  it('start() while signed out does NOT call the private listAccounts / getUsersByIds and does not error', async () => {
-    const { controller, oxy, sc } = makeHarness();
-    oxy.emitTokenChange(null); // cold boot: no bearer planted yet (no listeners registered pre-start)
-    sc.set(state([{ accountId: 'a1', sessionId: 's1' }], 'a1'));
-
-    controller.start();
-    await flush();
-
-    expect(oxy.listAccounts).not.toHaveBeenCalled();
-    expect(oxy.getUsersByIds).not.toHaveBeenCalled();
-    const snap = controller.getSnapshot();
-    expect(snap.error).toBeNull();
-    expect(snap.loading).toBe(false);
-    controller.destroy();
-  });
-
-  it('refresh() while signed out re-projects device-only and skips the network call', async () => {
-    const { controller, oxy } = makeHarness();
-    oxy.emitTokenChange(null);
-
-    await controller.refresh();
-
-    expect(oxy.listAccounts).not.toHaveBeenCalled();
-    const snap = controller.getSnapshot();
-    expect(snap.loading).toBe(false);
-    expect(snap.error).toBeNull();
-  });
-
-  it('start() while authenticated fetches the graph exactly once', async () => {
-    const { controller, oxy, sc } = makeHarness();
-    sc.set(state([{ accountId: 'a1', sessionId: 's1' }], 'a1'));
-    oxy.listAccounts.mockResolvedValue([graphNode('org1')]);
-    oxy.getUsersByIds.mockResolvedValue([user('a1'), user('org1')]);
-
-    controller.start();
-    await flush();
-
-    expect(oxy.listAccounts).toHaveBeenCalledTimes(1);
-    expect(controller.getSnapshot().accounts.map((r) => r.accountId)).toEqual(['a1', 'org1']);
-    controller.destroy();
-  });
-
-  it('fetches the graph once when the bearer is planted after a signed-out start', async () => {
-    const { controller, oxy } = makeHarness();
-    oxy.emitTokenChange(null);
-    controller.start();
-    await flush();
-    expect(oxy.listAccounts).not.toHaveBeenCalled();
-
-    // Cold-boot restore plants the token → onTokensChanged → single graph fetch.
-    oxy.emitTokenChange('access-token');
-    await flush();
-    expect(oxy.listAccounts).toHaveBeenCalledTimes(1);
-    controller.destroy();
-  });
-
-  it('drops the graph and re-projects device-only (no fetch) when the token is cleared', async () => {
-    const { controller, oxy, sc } = makeHarness();
-    sc.set(state([{ accountId: 'a1', sessionId: 's1' }], 'a1'));
-    oxy.listAccounts.mockResolvedValue([graphNode('org1')]);
-    oxy.getUsersByIds.mockResolvedValue([user('a1'), user('org1')]);
-    controller.start();
-    await flush();
-    expect(controller.getSnapshot().accounts.map((r) => r.accountId)).toEqual(['a1', 'org1']);
-
-    oxy.listAccounts.mockClear();
-    oxy.emitTokenChange(null); // a 401 cleared the bearer
-    await flush();
-
-    expect(oxy.listAccounts).not.toHaveBeenCalled();
-    // Graph-only org1 is gone; the device row survives.
-    expect(controller.getSnapshot().accounts.map((r) => r.accountId)).toEqual(['a1']);
-    controller.destroy();
-  });
-
-  it('does not loop when listAccounts rejects — at most one call per refresh, no re-trigger on device changes', async () => {
-    const { controller, oxy, sc } = makeHarness();
-    sc.set(state([{ accountId: 'a1', sessionId: 's1' }], 'a1'));
-    oxy.getUsersByIds.mockResolvedValue([user('a1')]);
-    oxy.listAccounts.mockRejectedValue(new Error('graph boom'));
-
-    controller.start();
-    await flush();
-    expect(oxy.listAccounts).toHaveBeenCalledTimes(1);
-    expect(controller.getSnapshot().error).toBe('graph boom');
-
-    // A subsequent device-state push must NOT re-trigger the graph fetch (auth
-    // edge unchanged → reconcileAuth is a no-op → no storm).
-    sc.set(state([{ accountId: 'a1', sessionId: 's1' }], 'a1', 2));
-    await flush();
-    expect(oxy.listAccounts).toHaveBeenCalledTimes(1);
-    // Device row still rendered despite the graph failure.
-    expect(controller.getSnapshot().accounts.map((r) => r.accountId)).toEqual(['a1']);
-    controller.destroy();
-  });
-});
-
-describe('AccountDialogController — switchTo (uniform switch)', () => {
-  it('uses SessionClient.switchAccount for an account already on the device', async () => {
-    const { controller, oxy, sc } = makeHarness();
-    sc.set(state([{ accountId: 'a1', sessionId: 's1' }, { accountId: 'a2', sessionId: 's2' }], 'a1'));
-    const switchSpy = jest.spyOn(sc, 'switchAccount').mockResolvedValue(undefined);
-    oxy.getUsersByIds.mockResolvedValue([user('a1'), user('a2')]);
-
-    await controller.switchTo('a2');
-
-    expect(switchSpy).toHaveBeenCalledWith('a2');
-    expect(oxy.switchToAccount).not.toHaveBeenCalled();
-    expect(controller.getSnapshot().switchingAccountId).toBeNull();
-  });
-
-  it('mints via oxyServices.switchToAccount + commitSession on first entry into a graph account', async () => {
-    const { controller, oxy, sc, commitSession } = makeHarness();
-    sc.set(state([{ accountId: 'a1', sessionId: 's1' }], 'a1'));
-    const switchSpy = jest.spyOn(sc, 'switchAccount').mockResolvedValue(undefined);
-    oxy.switchToAccount.mockResolvedValue({
-      sessionId: 'sess-org',
-      deviceId: 'device-1',
-      expiresAt: '2030-01-01T00:00:00Z',
-      accessToken: 'access-org',
-      user: user('org1'),
-    });
-    oxy.getUsersByIds.mockResolvedValue([user('a1'), user('org1')]);
-
-    await controller.switchTo('org1');
-
-    expect(oxy.switchToAccount).toHaveBeenCalledWith('org1');
-    expect(switchSpy).not.toHaveBeenCalled();
-    expect(commitSession).toHaveBeenCalledTimes(1);
-    expect(commitSession.mock.calls[0][0]).toMatchObject({ sessionId: 'sess-org', accessToken: 'access-org' });
-  });
-
-  it('commits a graph switch via the IN-PLACE commitSwitchedSession — never commitSession', async () => {
-    // An account switch must use commitSwitchedSession (in-place) and NEVER
-    // commitSession when both funnels are wired.
-    const oxy = makeOxy();
-    const sc = new TestSessionClient(host());
-    sc.set(state([{ accountId: 'a1', sessionId: 's1' }], 'a1'));
-    jest.spyOn(sc, 'switchAccount').mockResolvedValue(undefined);
-    oxy.switchToAccount.mockResolvedValue({
-      sessionId: 'sess-org',
-      deviceId: 'device-1',
-      expiresAt: '2030-01-01T00:00:00Z',
-      accessToken: 'access-org',
-      user: user('org1'),
-    });
-    const commitSession = jest.fn().mockResolvedValue(undefined);
-    const commitSwitchedSession = jest.fn().mockResolvedValue(undefined);
-    const controller = new AccountDialogController({
-      oxyServices: oxy as unknown as OxyServices,
-      sessionClient: sc,
-      clientId: 'oxy_dk_test',
-      commitSession,
-      commitSwitchedSession,
-    });
-
-    await controller.switchTo('org1');
-
-    expect(commitSwitchedSession).toHaveBeenCalledTimes(1);
-    expect(commitSwitchedSession.mock.calls[0][0]).toMatchObject({ sessionId: 'sess-org', accessToken: 'access-org' });
-    expect(commitSession).not.toHaveBeenCalled();
-  });
-
-  it('surfaces a failed switch as snapshot.error instead of silently no-op\'ing', async () => {
-    // PROBLEM 1: switching into an account the server refuses (e.g. a 403 for a
-    // personal-kind target) must tell the user why — the error is recorded on
-    // the snapshot, the switch does not throw, and switchingAccountId resets.
-    const { controller, oxy, sc } = makeHarness();
-    sc.set(state([{ accountId: 'a1', sessionId: 's1' }], 'a1'));
-    oxy.switchToAccount.mockRejectedValue(new Error('Cannot switch into a personal account'));
-
-    await expect(controller.switchTo('org1')).resolves.toBe(false);
-
-    const snap = controller.getSnapshot();
-    expect(snap.error).toBe('Cannot switch into a personal account');
-    expect(snap.switchingAccountId).toBeNull();
-  });
-
-  it('falls back to SessionClient.registerAndActivate when no commitSession is supplied', async () => {
-    const oxy = makeOxy();
-    const sc = new TestSessionClient(host());
-    sc.set(state([{ accountId: 'a1', sessionId: 's1' }], 'a1'));
-    const registerSpy = jest.spyOn(sc, 'registerAndActivate').mockResolvedValue(undefined);
-    oxy.switchToAccount.mockResolvedValue({
-      sessionId: 'sess-org',
-      deviceId: 'device-1',
-      expiresAt: '2030-01-01T00:00:00Z',
-      accessToken: 'access-org',
-      user: user('org1'),
-    });
-    const controller = new AccountDialogController({
-      oxyServices: oxy as unknown as OxyServices,
-      sessionClient: sc,
-      clientId: 'oxy_dk_test',
-    });
-
-    await controller.switchTo('org1');
-    expect(registerSpy).toHaveBeenCalledWith('org1');
-  });
-
-  it('ignores a concurrent switch while one is in flight', async () => {
-    const { controller, oxy, sc } = makeHarness();
-    sc.set(state([{ accountId: 'a1', sessionId: 's1' }, { accountId: 'a2', sessionId: 's2' }], 'a1'));
-    let release: () => void = () => undefined;
-    jest.spyOn(sc, 'switchAccount').mockImplementation(
-      () => new Promise<void>((resolve) => { release = resolve; }),
-    );
-
-    const first = controller.switchTo('a2');
-    expect(controller.getSnapshot().switchingAccountId).toBe('a2');
-    await controller.switchTo('a1'); // ignored — a switch is in flight
-    expect(sc.switchAccount).toHaveBeenCalledTimes(1);
-
-    release();
-    await first;
-  });
-
-  it('returns true when the switch succeeds even if the post-switch refresh fails', async () => {
-    const { controller, oxy, sc } = makeHarness();
-    sc.set(state([{ accountId: 'a1', sessionId: 's1' }, { accountId: 'a2', sessionId: 's2' }], 'a1'));
-    jest.spyOn(sc, 'switchAccount').mockResolvedValue(undefined);
-    oxy.listAccounts.mockRejectedValue(new Error('network down'));
-
-    await expect(controller.switchTo('a2')).resolves.toBe(true);
-
-    expect(controller.getSnapshot().error).toBe('network down');
-    expect(controller.getSnapshot().switchingAccountId).toBeNull();
   });
 });
 
@@ -891,7 +727,7 @@ describe('AccountDialogController — Commons availability (canOpenApp)', () => 
     oxy.pollCommonsSignIn.mockResolvedValue({ authorized: false, status: 'pending' });
     const controller = new AccountDialogController({
       oxyServices: oxy as unknown as OxyServices,
-      sessionClient: new TestSessionClient(host()),
+      sessionClient: new TestSessionClient(host().host),
       clientId: 'oxy_dk_test',
       pollIntervalMs: 1000,
       openUrl: opts.openUrl,
@@ -1015,7 +851,7 @@ describe('AccountDialogController — /auth-session socket (instant QR wake)', (
     const commitSession = jest.fn().mockResolvedValue(undefined);
     const controller = new AccountDialogController({
       oxyServices: oxy as unknown as OxyServices,
-      sessionClient: new TestSessionClient(host()),
+      sessionClient: new TestSessionClient(host().host),
       clientId: 'oxy_dk_test',
       commitSession,
       socketFactory: factory as unknown as SocketIOFactory,
@@ -1134,7 +970,7 @@ function makeDeliveryHarness(opts: {
   if (opts.authenticated === false) oxy.emitTokenChange(null);
   const controller = new AccountDialogController({
     oxyServices: oxy as unknown as OxyServices,
-    sessionClient: new TestSessionClient(host()),
+    sessionClient: new TestSessionClient(host().host),
     clientId: 'oxy_dk_test',
     // Every real consumer wires the provider's commit funnel; without it the
     // controller falls back to `SessionClient.registerAndActivate`, which opens
@@ -1428,7 +1264,7 @@ describe('AccountDialogController — automatic delivery selection (#691 phase 5
     oxy.deliverCommonsSignIn.mockResolvedValue({ delivered: true, targets: 1 });
     const controller = new AccountDialogController({
       oxyServices: oxy as unknown as OxyServices,
-      sessionClient: new TestSessionClient(host()),
+      sessionClient: new TestSessionClient(host().host),
       clientId: 'oxy_dk_test',
       pollIntervalMs: 1000,
       platform: 'desktop',
@@ -1676,7 +1512,7 @@ describe('AccountDialogController — cancellation converges (#691 phase 5)', ()
       oxy.pollCommonsSignIn.mockResolvedValue({ authorized: false, status: 'pending' });
       const controller = new AccountDialogController({
         oxyServices: oxy as unknown as OxyServices,
-        sessionClient: new TestSessionClient(host()),
+        sessionClient: new TestSessionClient(host().host),
         clientId: 'oxy_dk_test',
         pollIntervalMs: 1000,
         platform: 'desktop',
@@ -1792,7 +1628,7 @@ describe('AccountDialogController — cancellation converges (#691 phase 5)', ()
       });
       const controller = new AccountDialogController({
         oxyServices: oxy as unknown as OxyServices,
-        sessionClient: new TestSessionClient(host()),
+        sessionClient: new TestSessionClient(host().host),
         clientId: 'oxy_dk_test',
         pollIntervalMs: 1000,
         platform: 'desktop',
@@ -1986,45 +1822,50 @@ describe('AccountDialogController — the directory (ADR 0002)', () => {
     expect(controller.getSnapshot().activatingContextId).toBeNull();
   });
 
-  it('keeps rendering the flat list when the directory read fails', async () => {
-    const oxy = makeOxy();
-    const sc = new TestSessionClient({
-      makeRequest: jest.fn(async (_method: string, url: string) => {
-        if (url === '/session/device/directory') throw new Error('boom');
-        return undefined;
-      }),
-      getBaseURL: () => 'http://test.invalid',
-      getAccessToken: () => 'token',
-      getDeviceCredential: () => null,
-      onTokensChanged: () => () => undefined,
-      setTokens: jest.fn(),
-      getCurrentAccountId: () => null,
-    });
-    const controller = createAccountDialogController({
-      oxyServices: oxy as unknown as OxyServices,
-      sessionClient: sc,
-      clientId: 'oxy_dk_test',
-      commitSession: jest.fn().mockResolvedValue(undefined),
-      onSignedIn: jest.fn(),
-    });
-    sc.set(state([{ accountId: 'a1', sessionId: 's1' }], 'a1'));
-    oxy.listAccounts.mockResolvedValue([]);
-    oxy.getUsersByIds.mockResolvedValue([user('a1')]);
+  it('keeps the directory it already holds when a later read fails, and says so', async () => {
+    const warnSpy = jest.spyOn(logger, 'warn').mockImplementation(() => undefined);
+    const { controller, failDirectory } = makeHarness();
 
     await controller.refresh();
+    expect(controller.getSnapshot().directory?.revision).toBe(9_000);
 
-    // Degraded, not blank: a directory outage must not take the switcher down
-    // while the compatibility lane can still answer.
+    failDirectory(new Error('boom'));
+    await controller.refresh();
+
+    // Degraded, not blank. A transient outage must not empty a switcher the
+    // user is looking at — but it must not pretend either, so the error is on
+    // the snapshot for the surface to render.
     const snap = controller.getSnapshot();
-    expect(snap.directory).toBeNull();
-    expect(snap.accounts.map((r) => r.accountId)).toEqual(['a1']);
-    expect(snap.error).toBeNull();
+    expect(snap.directory?.revision).toBe(9_000);
+    expect(snap.error).toBe('boom');
+    expect(snap.loading).toBe(false);
+    expect(warnSpy).toHaveBeenCalledWith(
+      '[AccountDialogController] directory refresh failed',
+      { component: 'AccountDialogController' },
+      expect.any(Error),
+    );
+    warnSpy.mockRestore();
+  });
+
+  it('spins only while there is nothing to show', async () => {
+    const { controller } = makeHarness();
+    const loading: boolean[] = [];
+    controller.subscribe((s) => loading.push(s.loading));
+
+    // First read: nothing held, so the switcher has nothing to render.
+    await controller.refresh();
+    expect(loading).toContain(true);
+    expect(controller.getSnapshot().loading).toBe(false);
+
+    // Second read, behind a directory already on screen: refreshed in place,
+    // never blanked back to a spinner.
+    loading.length = 0;
+    await controller.refresh();
+    expect(loading).not.toContain(true);
   });
 
   it('removes ONE route to a shared account, naming the pair and not the account', async () => {
-    const { controller, oxy, urls, bodies } = directoryHarness('ctx-nate-org');
-    oxy.listAccounts.mockResolvedValue([]);
-    oxy.getUsersByIds.mockResolvedValue([]);
+    const { controller, urls, bodies } = directoryHarness('ctx-nate-org');
     await controller.refresh();
 
     const inFlight: Array<string | null> = [];
@@ -2043,9 +1884,7 @@ describe('AccountDialogController — the directory (ADR 0002)', () => {
   });
 
   it('removes ONE PERSON through the principal id, not through their contexts', async () => {
-    const { controller, oxy, urls, bodies } = directoryHarness('ctx-nate-org');
-    oxy.listAccounts.mockResolvedValue([]);
-    oxy.getUsersByIds.mockResolvedValue([]);
+    const { controller, urls, bodies } = directoryHarness('ctx-nate-org');
     await controller.refresh();
 
     const inFlight: Array<string | null> = [];
@@ -2064,9 +1903,7 @@ describe('AccountDialogController — the directory (ADR 0002)', () => {
   });
 
   it('refuses a second removal while one is in flight, of either kind', async () => {
-    const { controller, oxy, sc } = directoryHarness('ctx-nate-org');
-    oxy.listAccounts.mockResolvedValue([]);
-    oxy.getUsersByIds.mockResolvedValue([]);
+    const { controller, sc } = directoryHarness('ctx-nate-org');
     await controller.refresh();
     let release: () => void = () => undefined;
     jest.spyOn(sc, 'signOutContext').mockImplementation(
@@ -2085,9 +1922,7 @@ describe('AccountDialogController — the directory (ADR 0002)', () => {
   });
 
   it('surfaces a failed removal without wedging the flag', async () => {
-    const { controller, oxy, sc } = directoryHarness('ctx-nate-org');
-    oxy.listAccounts.mockResolvedValue([]);
-    oxy.getUsersByIds.mockResolvedValue([]);
+    const { controller, sc } = directoryHarness('ctx-nate-org');
     await controller.refresh();
     // A context id is not stable across a removal, so a stale one is ordinary:
     // the server refuses and the next read simply does not offer it.
