+++ packages/services/src/ui/hooks/accountDialogSnapshot.ts
@@ -14,8 +14,6 @@ import type { AccountDialogSnapshot } from '@oxyhq/core';
  */
 export const EMPTY_ACCOUNT_DIALOG_SNAPSHOT: AccountDialogSnapshot = {
   view: 'accounts',
-  accounts: [],
-  activeAccountId: null,
   directory: null,
   activeContext: null,
   activatingContextId: null,
@@ -23,7 +21,6 @@ export const EMPTY_ACCOUNT_DIALOG_SNAPSHOT: AccountDialogSnapshot = {
   removingPrincipalId: null,
   loading: false,
   error: null,
-  switchingAccountId: null,
   signIn: {
     phase: 'idle',
     authorizeCode: null,
